Version 2.1.4
- added support for blue STM32F103ZET6 board
- supported by classes GxIO_STM32F103ZET6_P16 and GxIO_STM32F103ZET6_FSMC
- fixed pushColors methods
- implemented writeRect and pushRect methods (these were empty)

+++ src/GxTFT.h
@@ -288,7 +288,7 @@ class GxTFT : public Print
              pushColor(uint16_t color),
              pushColor(uint16_t color, uint16_t len),
 
-             pushColors(uint16_t *data, uint8_t len),
+             pushColors(uint16_t *data, uint32_t len),
              pushColors(uint8_t  *data, uint32_t len),
 
              fillScreen(uint16_t color),
